// frontend/src/interfaces/product.ts
export interface ProductInterface {
    _id: string
    name: string
    price: number
    image: string
}

export interface ProductDetailsInterface {
    _id: string
    name: string
    release: string
    developer: string
    publisher: string
    description: string
    image: string
    price: number
}